test: refactor test functions
